pyproject.toml
[tool.poetry]
name = "difference_between_file"
version = "0.1.0"
description = ""
authors = ["a-paktusov@yandex.ru <83519069946>"]

[tool.poetry.dependencies]
python = "^3.10"
python-docx = "^0.8.11"
fastapi = "^0.85.1"
uvicorn = "^0.18.3"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
